Refactor the front end code to properly follow the conventions (#99)
