+++ client/src/components/Wall/NavBar.js
@@ -6,8 +6,6 @@ import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
 import InputBase from '@material-ui/core/InputBase';
 import Badge from '@material-ui/core/Badge';
-import MenuItem from '@material-ui/core/MenuItem';
-import Menu from '@material-ui/core/Menu';
 import SearchIcon from '@material-ui/icons/Search';
 import NotificationsIcon from '@material-ui/icons/Notifications';
 import MoreIcon from '@material-ui/icons/MoreVert';
@@ -77,18 +75,14 @@ const useStyles = makeStyles(theme => ({
 
 export default function PrimarySearchAppBar() {
     const classes = useStyles();
-    const [anchorEl, setAnchorEl] = React.useState(null);
-    const [mobileMoreAnchorEl, setMobileMoreAnchorEl] = React.useState(null);
     const [numberOfNotifications,setNumberOfNotifications]=React.useState(0);
-    const isMenuOpen = Boolean(anchorEl);
-    const isMobileMenuOpen = Boolean(mobileMoreAnchorEl);
 
 
     useEffect(()=>{
         request.put("http://localhost:8000/api/notification")
             .query({email: localStorage.getItem("email")})
             .end((err,res) =>{
-                if(res.status==222){
+                if(res.status===222){
                     console.log("notification set");
                    setNumberOfNotifications(1);
                 }
@@ -101,65 +95,18 @@ export default function PrimarySearchAppBar() {
         request.put("http://localhost:8000/api/notification/remove")
             .query({email: localStorage.getItem("email")})
             .end((err,res) => {
-                if (res.status == 224) {
+                if (res.status === 224) {
                     console.log("removed notification");
                 }
             });
         window.location.reload(false);
     }
 
-
-    const handleProfileMenuOpen = event => {
-        setAnchorEl(event.currentTarget);
-    };
-
-    const handleMobileMenuClose = () => {
-        setMobileMoreAnchorEl(null);
-    };
-
-
-    const handleMobileMenuOpen = event => {
-        setMobileMoreAnchorEl(event.currentTarget);
-    };
-
     const menuId = 'primary-search-account-menu';
 
     const mobileMenuId = 'primary-search-account-menu-mobile';
-    // const renderMobileMenu = (
-    //     <Menu
-    //         anchorEl={mobileMoreAnchorEl}
-    //         anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-    //         id={mobileMenuId}
-    //         keepMounted
-    //         transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-    //         open={isMobileMenuOpen}
-    //         onClose={handleMobileMenuClose}
-    //     >
-    //         <MenuItem>
-    //             <IconButton aria-label="show 2 new notifications" color="inherit" onClick={console.log("PP")}>
-    //                 {/*TODO:Use the real number of notification*/}
-    //                 <Badge badgeContent={numberOfNotifications} color="secondary">
-    //                     <NotificationsIcon />
-    //                 </Badge>
-    //             </IconButton>
-    //             <p>Notifications</p>
-    //         </MenuItem>
-    //         <MenuItem onClick={handleProfileMenuOpen}>
-    //             <IconButton
-    //                 aria-label="account of current user"
-    //                 aria-controls="primary-search-account-menu"
-    //                 aria-haspopup="true"
-    //                 color="inherit"
-    //             >
-    //                 {/*TODO: Use Picture for avatar */}
-    //                 <Avatar className={classes.orange}>N</Avatar>
-    //             </IconButton>
-    //             <p>Profile</p>
-    //         </MenuItem>
-    //     </Menu>
-    // );
 
-    function RedirectToWall(e){
+    function redirectToWall(e){
         console.log("Redirecting");
         history.push('/wall');
         window.location.reload(false);
@@ -179,7 +126,7 @@ export default function PrimarySearchAppBar() {
                         aria-label="open drawer"
                     >
                     </IconButton>
-                    <Typography className={classes.title} variant="h4" noWrap onClick={RedirectToWall}>
+                    <Typography className={classes.title} variant="h4" noWrap onClick={redirectToWall}>
                         Kandid
                     </Typography>
                     <div className={classes.search}>
@@ -209,7 +156,6 @@ export default function PrimarySearchAppBar() {
                             aria-label="account of current user"
                             aria-controls={menuId}
                             aria-haspopup="true"
-                            onClick={handleProfileMenuOpen}
                             color="inherit"
                         >
                             <Avatar className={classes.orange}>N</Avatar>
@@ -220,7 +166,6 @@ export default function PrimarySearchAppBar() {
                             aria-label="show more"
                             aria-controls={mobileMenuId}
                             aria-haspopup="true"
-                            onClick={handleMobileMenuOpen}
                             color="inherit"
                         >
                             <MoreIcon />
@@ -228,7 +173,6 @@ export default function PrimarySearchAppBar() {
                     </div>
                 </Toolbar>
             </AppBar>
-            {/*{renderMobileMenu}*/}
         </div>
     );
 }
